// src/app/core/api/model/faq-update-dto.ts
/**
 * E-commerce platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.4.22
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FaqUpdateDto { 
    question?: string;
    visible?: boolean;
    answer?: string;
    question_en?: string;
    answer_en?: string;
    category?: string;
    order?: number;
    is_active?: boolean;
}